minor refactoring
- moved EditPatientDetails form to a separate file
- added the test file
